fix(ci): macOS build — move libc to cfg(unix), cfg-gate Linux-only imports
The --color-tune PR (#399) introduced memstat.rs which uses libc on macOS
for mach_task_basic_info RSS sampling. But libc was declared as
cfg(target_os = "linux") in Cargo.toml, so macOS builds failed with
'unresolved import libc'.

Fixes:
1. Cargo.toml: move libc = "0.2" from cfg(target_os = "linux") to
   cfg(unix) so Linux + macOS + BSDs all get it. Linux target block
   left as empty placeholder with explanatory comment.
2. memstat.rs: cfg-gate 'use std::io::Read' to cfg(target_os = "linux")
   only — it's only used in the /proc parsing branch. Unconditional
   import triggered unused-import warning on macOS.
3. memstat.rs: remove redundant 'TASK_BASIC_INFO as u32' cast. The
   constant is already task_flavor_t (u32) — cast would trigger clippy
   useless_conversion on macOS.
4. main.rs: cfg-gate 'use crate::terminal::restore_terminal_best_effort'
   to cfg(target_os = "linux") only — its only caller is
   spawn_kill9_terminal_guard() which is cfg(target_os = "linux").
5. event_loop.rs: cfg-gate 'let mut last_ctxt_switches' to
   cfg(target_os = "linux") only — pre-existing warning that surfaced
   on macOS where /proc context-switch sampling is skipped.

Verification (Linux, cannot cross-compile to macOS locally):
- cargo build --locked clean (CI uses --locked)
- cargo clippy --all-targets -- -D warnings clean
- cargo fmt --all -- --check clean
- cargo test --bin cosmostrix: 837 pass, 0 fail
- macOS code path hand-reviewed: libc 0.2 exposes all symbols used
  (mach_task_self, task_info, task_basic_info, TASK_BASIC_INFO,
  task_info_t, natural_t, mach_msg_type_number_t, KERN_SUCCESS, c_int)

+++ src/memstat.rs
@@ -25,6 +25,7 @@
 //! process footprint", not as a precise allocator accounting. For allocator
 //! accounting, run under `valgrind --tool=massif` or `heaptrack` separately.
 
+#[cfg(target_os = "linux")]
 use std::io::Read;
 
 /// Sample the current process RSS in kibibytes (KiB), if available.
@@ -105,7 +106,7 @@ fn macos_rss_kb() -> Option<u64> {
             as libc::mach_msg_type_number_t;
         let kr: c_int = task_info(
             mach_task_self(),
-            TASK_BASIC_INFO as u32,
+            TASK_BASIC_INFO,
             &mut info as *mut _ as task_info_t,
             &mut count,
         );
